Separate UI components
New function components JenMenu and JenSlider
Move container to MediaController
