+++ Lux/lux_react/src/components/MediaController.js
@@ -1,12 +1,13 @@
 import React, { useState } from 'react';
-import useEmscripten from './useEmscripten';
+import Paper from '@mui/material/Paper';
+//import useEmscripten from './useEmscripten';
 import { ButtonGroup, Button, Tooltip } from '@mui/material';
 import RestartIcon from '@mui/icons-material/Replay'; // This is just an example icon for "restart"
 import FrameIcon from '@mui/icons-material/SkipNext'; // This is an example icon for "advance one frame"
 import PlayPauseIcon from '@mui/icons-material/PlayArrow'; // This is an example icon for "play"
 import PauseIcon from '@mui/icons-material/Pause'; // This is an example icon for "pause"
 
-const MediaController = () => {
+function MediaController( { panelSize } ) { 
   const [isRunning, setIsRunning] = useState(true);
 //  module = useEmscripten();
 
@@ -31,34 +32,45 @@ const MediaController = () => {
   }
 
   return (
-    <ButtonGroup 
-      variant="contained"
-      sx={{
-        maxHeight: '40px',
-        '& .MuiButton-root': {
-          width: '40px',
-          height: '40px'
-        }
-      }}
+    <Paper 
+    elevation={3}
+    sx={{
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+      width: panelSize,
+      height: 60,  
+    }}
     >
-      <Tooltip title="Restart">
-        <Button onClick={handleRestart}>
-          <RestartIcon />
-        </Button>
-      </Tooltip>
+      <ButtonGroup 
+        variant="contained"
+        sx={{
+          maxHeight: '40px',
+          '& .MuiButton-root': {
+            width: '40px',
+            height: '40px'
+          }
+        }}
+      >
+        <Tooltip title="Restart">
+          <Button onClick={handleRestart}>
+            <RestartIcon />
+          </Button>
+        </Tooltip>
 
-      <Tooltip title="Advance One Frame">
-        <Button onClick={handleAdvance}>
-          <FrameIcon />
-        </Button>
-      </Tooltip>
+        <Tooltip title="Advance One Frame">
+          <Button onClick={handleAdvance}>
+            <FrameIcon />
+          </Button>
+        </Tooltip>
 
-      <Tooltip title={isRunning ? "Pause" : "Run"}>
-      <Button onClick={handleRunPause}>
-          {isRunning ? <PauseIcon /> : <PlayPauseIcon />}
-        </Button>
-      </Tooltip>
-    </ButtonGroup>
+        <Tooltip title={isRunning ? "Pause" : "Run"}>
+        <Button onClick={handleRunPause}>
+            {isRunning ? <PauseIcon /> : <PlayPauseIcon />}
+          </Button>
+        </Tooltip>
+      </ButtonGroup>
+    </Paper>
   );
 };
